fix bug base64

+++ src/main/java/com/plataformae/ws/db/repository/IDepartamentoRepository.java
@@ -3,5 +3,5 @@ package com.plataformae.ws.db.repository;
 import com.plataformae.ws.db.entity.Departamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDepartamentoRepository extends JpaRepository<Departamento, Integer> {
+public interface IDepartamentoRepository extends JpaRepository<Departamento, Long> {
 }
